+++ inst/tinytest/test_reaggregate_interval_counts.R
@@ -1,19 +1,19 @@
 # without weights
-lower_bounds <- c(1, 2, NA,  2, 1, NA, 4, 6)
-upper_bounds <- c(3, 3,  1, NA, 3, NA, 6, Inf)
-counts <- c(1, 1,  1,  1, 1,  1, NA_real_, 10)
+lower_bounds <- c(1, 2, 1,        4,   6)
+upper_bounds <- c(3, 3, 3,        6, Inf)
+counts       <- c(1, 1, 1, NA_real_,  10)
 max_upper <- 100
 breaks <- c(0, 1, 2, 6)
 
 expected <- data.frame(
     interval = factor(
-        c("[0, 1)", "[1, 2)", "[2, 6)", "[6, Inf)", NA_character_),
+        c("[0, 1)", "[1, 2)", "[2, 6)", "[6, Inf)"),
         levels = c("[0, 1)", "[1, 2)", "[2, 6)", "[6, Inf)"),
         ordered = TRUE
     ),
-    lower_bound = c(0, 1, 2, 6, NA),
-    upper_bound = c(1, 2, 6, Inf, NA),
-    count = c(0, 1, NA, 10, 3)
+    lower_bound = c(0, 1, 2, 6),
+    upper_bound = c(1, 2, 6, Inf),
+    count = c(0, 1, NA, 10)
 )
 
 expect_equal(
@@ -32,13 +32,13 @@ weights <- c(1, 3, rep.int(1, 98))
 
 expected <- data.frame(
     interval = factor(
-        c("[0, 1)", "[1, 2)", "[2, 6)", "[6, Inf)", NA_character_),
+        c("[0, 1)", "[1, 2)", "[2, 6)", "[6, Inf)"),
         levels = c("[0, 1)", "[1, 2)", "[2, 6)", "[6, Inf)"),
         ordered = TRUE
     ),
-    lower_bound = c(0, 1, 2, 6, NA),
-    upper_bound = c(1, 2, 6, Inf, NA),
-    count = c(0, 1.5, NA, 10, 3)
+    lower_bound = c(0, 1, 2, 6),
+    upper_bound = c(1, 2, 6, Inf),
+    count = c(0, 1.5, NA, 10)
 )
 
 expect_equal(