+++ inst/tinytest/test_split_interval_counts.R
@@ -1,38 +1,40 @@
 # without weights
 result_1 <- split_interval_counts(
-    lower_bounds = c(1, 2, NA,  2, 1, NA, 4),
-    upper_bounds = c(3, 3,  1, NA, 3, NA, 6),
-    counts =       c(1, 1,  1,  1, 1,  1, NA_real_)
+    lower_bounds = c(1, 2, 1, 4),
+    upper_bounds = c(3, 3, 3, 6),
+    counts =       c(1, 1, 1, NA_real_)
 )
 
 expected_1 <- data.frame(
     age = c(
         1, 2,
         2,
-        NA_integer_,
-        NA_integer_,
         1, 2,
-        NA_integer_,
         4, 5
     ),
     count = c(
         0.5, 0.5,
         1,
-        1,
-        1,
         0.5, 0.5,
-        1,
         NA_real_, NA_real_
     )
 )
 
 expect_equal(result_1, expected_1)
+expect_identical(
+    result_1,
+    ageutils:::split_interval_counts_r(
+        lower_bounds = c(1, 2, 1, 4),
+        upper_bounds = c(3, 3, 3, 6),
+        counts =       c(1, 1, 1, NA_real_)
+    )
+)
 
 # with weights
 result_2 <- split_interval_counts(
-    lower_bounds = c(1, 2, NA,  2, 1, NA, 4),
-    upper_bounds = c(3, 3,  1, NA, 3, NA, 6),
-    counts =       c(1, 1,  1,  1, 1,  1, NA_real_),
+    lower_bounds = c(1, 2, 1, 4),
+    upper_bounds = c(3, 3, 3, 6),
+    counts =       c(1, 1, 1, NA_real_),
     weights = c(1, 3, 1, 1, 1, 1),
     max_upper = 6
 )
@@ -41,24 +43,28 @@ expected_2 <- data.frame(
     age = c(
         1, 2,
         2,
-        NA_integer_,
-        NA_integer_,
         1, 2,
-        NA_integer_,
         4, 5
     ),
     count = c(
         0.75, 0.25,
         1,
-        1,
-        1,
         0.75, 0.25,
-        1,
         NA_real_, NA_real_
     )
 )
 
 expect_equal(result_2, expected_2)
+expect_identical(
+    result_2,
+    ageutils:::split_interval_counts_r(
+        lower_bounds = c(1, 2, 1, 4),
+        upper_bounds = c(3, 3, 3, 6),
+        counts =       c(1, 1, 1, NA_real_),
+        weights = c(1, 3, 1, 1, 1, 1),
+        max_upper = 6
+    )
+)
 
 # zero weights
 result_3 <- split_interval_counts(
@@ -73,6 +79,17 @@ expected_3 <- data.frame(
     count = c(rep.int(1L, 5L), rep.int(0L, 5L))
 )
 expect_equal(result_3, expected_3)
+expect_identical(
+    result_3,
+    ageutils:::split_interval_counts_r(
+        lower_bounds = c(1L, 6L),
+        upper_bounds = c(6L, 11L),
+        counts = c(5L, 5L),
+        max_upper = 11,
+        weights = c(rep.int(1L, 6L), rep.int(0L, 5L))
+    )
+)
+
 
 # input checking
 expect_error(
@@ -81,8 +98,6 @@ expect_error(
         upper_bounds = c(3, 3,  1, NA, 3, NA, 6),
         counts =       c(1, 1,  1,  1, 1,  1, NA_real_)
     ),
-    "`lower_bounds` must be numeric.",
-    fixed = TRUE
 )
 
 expect_error(
@@ -90,9 +105,7 @@ expect_error(
         lower_bounds = c(1, 2, NA,  2, 1, NA, 4),
         upper_bounds = as.character(c(3, 3,  1, NA, 3, NA, 6)),
         counts =       c(1, 1,  1,  1, 1,  1, NA_real_)
-    ),
-    "`upper_bounds` must be numeric.",
-    fixed = TRUE
+    )
 )
 
 expect_error(
@@ -100,9 +113,7 @@ expect_error(
         lower_bounds = c(1, 2, NA,  2, 1, NA, 4),
         upper_bounds = c(3, 3,  1, NA, 3, NA, 6),
         counts =       as.character(c(1, 1,  1,  1, 1,  1, NA_real_))
-    ),
-    "`counts` must be numeric.",
-    fixed = TRUE
+    )
 )
 
 expect_error(
@@ -111,9 +122,7 @@ expect_error(
         upper_bounds = c(3, 3,  1, NA, 3, NA, 6),
         counts =       c(1, 1,  1,  1, 1,  1, NA_real_),
         max_upper = "100"
-    ),
-    "`max_upper` must be an integer of length 1.",
-    fixed = TRUE
+    )
 )
 
 expect_error(
@@ -122,27 +131,14 @@ expect_error(
         upper_bounds = c(3, 3,  1, NA, 3, NA, 6),
         counts =       c(1, 1,  1,  1, 1,  1, NA_real_),
         max_upper = 100:101
-    ),
-    "`max_upper` must be an integer of length 1.",
-    fixed = TRUE
+    )
 )
 
 expect_error(
     split_interval_counts(
-        lower_bounds = c(1, 2, NA,  2, 1, NA, 4),
-        upper_bounds = c(3, 3,  1, NA, 3, NA, 6),
-        counts =       c(1, 1,  1,  1, 1,  1, NA_real_),
-        max_upper = 201
-    ),
-    "`max_upper` must be less than or equal to 200.",
-    fixed = TRUE
-)
-
-expect_error(
-    split_interval_counts(
-        lower_bounds = c(1, 2, NA,  2, 1, NA, 7),
-        upper_bounds = c(3, 3,  1, NA, 3, NA, 6),
-        counts =       c(1, 1,  1,  1, 1,  1, NA_real_)
+        lower_bounds = c(1, 2, 1, 7),
+        upper_bounds = c(3, 3, 3, 6),
+        counts =       c(1, 1, 1, NA_real_)
     ),
     "`lower_bounds` must be less than `upper_bounds`.",
     fixed = TRUE
@@ -161,16 +157,6 @@ expect_error(
 expect_error(
     split_interval_counts(
         lower_bounds = c(1, 2, NA,  2, 1, NA, 4),
-        upper_bounds = c(3, 3,  1, NA, 3, NA, 101),
-        counts =       c(1, 1,  1,  1, 1,  1, NA_real_)
-    ),
-    "`upper_bounds` can not be greater than `max_upper` unless infinite.",
-    fixed = TRUE
-)
-
-expect_error(
-    split_interval_counts(
-        lower_bounds = c(1, 2, NA,  2, 1, NA, 4),
         upper_bounds = c(3, 3,  1, NA, 3, NA, 6),
         counts =       c(1, 1,  1,  1, 1,  1)
     ),
@@ -180,9 +166,9 @@ expect_error(
 
 expect_error(
     split_interval_counts(
-        lower_bounds = c(1, 2, NA,  2, 1, NA, 4),
-        upper_bounds = c(3, 3,  1, NA, 3, NA, 6),
-        counts =       c(1, 1,  1,  1, 1,  1, NA_real_),
+        lower_bounds = c(1, 2, 1, 4),
+        upper_bounds = c(3, 3, 3, 6),
+        counts =       c(1, 1, 1, NA_real_),
         max_upper = 50,
         weights = 201
     ),
@@ -192,47 +178,44 @@ expect_error(
 
 expect_error(
     split_interval_counts(
-        lower_bounds = c(1, 2, NA,  2, 1, NA, 4),
-        upper_bounds = c(3, 3,  1, NA, 3, NA, 6),
-        counts =       c(1, 1,  1,  1, 1,  1, NA_real_),
+        lower_bounds = c(1, 2, 1, 4),
+        upper_bounds = c(3, 3, 3, 6),
+        counts =       c(1, 1, 1, NA_real_),
         weights = ""
-    ),
-    "`weights` must be numeric.",
-    fixed = TRUE
+    )
 )
+
 max_upper <- 50L
 weights <- seq_len(max_upper)
 weights[1] <- - weights[1]
 expect_error(
     split_interval_counts(
-        lower_bounds = c(1, 2, NA,  2, 1, NA, 4),
-        upper_bounds = c(3, 3,  1, NA, 3, NA, 6),
-        counts =       c(1, 1,  1,  1, 1,  1, NA_real_),
+        lower_bounds = c(1, 2, 1, 4),
+        upper_bounds = c(3, 3, 3, 6),
+        counts =       c(1, 1, 1, NA_real_),
         max_upper = max_upper,
         weights = weights
-    ),
-    "`weights` must be positive and not missing (NA).",
-    fixed = TRUE
+    )
 )
 
 weights[1] <- NA_integer_
 expect_error(
     split_interval_counts(
-        lower_bounds = c(1, 2, NA,  2, 1, NA, 4),
-        upper_bounds = c(3, 3,  1, NA, 3, NA, 6),
-        counts =       c(1, 1,  1,  1, 1,  1, NA_real_),
+        lower_bounds = c(1, 2, 1, 4),
+        upper_bounds = c(3, 3, 3, 6),
+        counts =       c(1, 1, 1, NA_real_),
         max_upper = max_upper,
         weights = weights
-    ),
-    "`weights` must be positive and not missing (NA).",
-    fixed = TRUE
+    )
 )
 
 # check success
-expect_silent(
+expect_warning(
     split_interval_counts(
-        lower_bounds = c(1, 2, NA,  2, 1, NA, 4),
-        upper_bounds = c(3, 3,  1, NA, 3, NA, Inf),
-        counts =       c(1, 1,  1,  1, 1,  1, NA_real_)
-    )
+        lower_bounds = c(1, 2, 1, 4),
+        upper_bounds = c(3, 3, 3, Inf),
+        counts =       c(1, 1, 1, NA_real_)
+    ),
+    "`upper_bounds` greater than `max_upper` (100) have been replaced prior to splitting.",
+    fixed = TRUE
 )