+++ NEWS.md
@@ -1,3 +1,22 @@
+# ageutils 0.0.2
+
+* `aggregate_age_counts()` will now only return a row corresponding to NA ages
+  if they were present in the input data. Previously an NA-associated row would
+  always be returned even if it's count was 0. Due to this change
+  `reaggregate_interval_counts()` will now never return an NA-associated row.
+
+* `split_interval_counts()` now matches the documentation and disallows missing
+  (NA) bounds.
+
+* `breaks_to_interval()` and `cut_ages()` both gain an argument, `max_upper`
+  which allows users to explicitly set the maximum upper bound.
+  
+* New function `reaggregate_interval_rates()`.
+
+* We now use [markdown](https://cran.r-project.org/package=markdown) as an
+  lighter alternative to [rmarkdown](https://cran.r-project.org/package=rmarkdown)
+  for the vignette.
+
 # ageutils 0.0.1
 
 Initial release of `ageutils` which provides a collection of efficient functions